@@ -1,7 +1,7 @@
 package eu.tutorials.shaproject
 
 import android.content.Context
-import android.support.v7.widget.RecyclerView
+import androidx.recyclerview.widget.RecyclerView
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
@@ -15,7 +15,8 @@ class Adapter_coursesscreen (val context: Context, val items: List<CoursesX>) :
     }
     var courseClickListener: CourseClickListener? = null
 
-    class ViewHolder(itemView:View,private val items: List<CoursesX>,private val courseClickListener:CourseClickListener ):RecyclerView.ViewHolder(itemView){
+    class ViewHolder(itemView:View,private val items: List<CoursesX>,private val courseClickListener:CourseClickListener ):
+        RecyclerView.ViewHolder(itemView){
         var course_name: TextView = itemView.course_1
 
         init {
