@@ -1,21 +1,16 @@
 package eu.tutorials.shaproject
 
-import android.app.DatePickerDialog
-import android.support.v7.app.AppCompatActivity
+import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
-import android.view.View
-import android.widget.Toast
 import kotlinx.android.synthetic.main.activity_main.*
-import java.text.SimpleDateFormat
-import java.util.*
 import android.content.Intent
 
 class MainActivity : AppCompatActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         setContentView(R.layout.activity_main)
-        intro_scree.alpha =0f
-        intro_scree.animate().setDuration(2500).alpha(1f).withEndAction(){
+        imgIntroScreen.alpha =0f
+        imgIntroScreen.animate().setDuration(2500).alpha(1f).withEndAction(){
             val i=Intent(this,Login::class.java)
             startActivity(i)
             finish()
