@@ -2,9 +2,9 @@ package eu.tutorials.shaproject
 
 import android.content.Intent
 import android.graphics.drawable.Drawable
-import android.support.v7.app.AppCompatActivity
+import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
-import android.support.v4.content.ContextCompat
+import androidx.core.content.ContextCompat
 import android.util.Log
 import android.widget.Toast
 import kotlinx.android.synthetic.main.activity_coursescreen.*
