@@ -1,6 +1,6 @@
 package eu.tutorials.shaproject
 
-import android.support.v7.app.AppCompatActivity
+import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
 
 class student_details : AppCompatActivity() {
